Update RNotebook.Rmd
Lauren Brese contributed to the team by finding the coding needed to try and run the LED board experiment and by bringing the Raspberry Pi computer needed to complete the experiment.  The package gpiozero was first downloaded onto the Raspberry Pi to make sure that the code could run and once the package was added to the Raspberry Pi, the code could be imported.  The coding began with 

from gpiozero import LEDBoard
from time import sleep
from signal import pause

leds = LEDBoard(40, 45, 50, 55, 60)

leds.on()
sleep(1)
leds.off()
sleep(1)
leds.value = (1, 0, 1, 0, 1)
sleep(1)
leds.blink()

pause()

After the code was added to the terminal, the code was run, only to find that there was an error in the coding.  The coding was changed five times, thinking that it was the numbers for the lights that were the issue, and after each change, the code was run to see what the error was.  However, after the fifth run, the terminal froze and no edits could be made.  Multiple attempts were made to try and click on different programs on the Raspberry Pi, but it seemed that the system had crashed and nothing was working.  The Raspberry Pi was given a hard reboot twice, but both times the system would only get up to about halfway before an error message was shown that said "Can not sync".  After the failed hard reboots, a picture was taken of the device set up and then disassembled.

Overall, the setup and lead up to the experiment was a success, but unfortunately, the coding and actual results of the experiment could not be obtained.